refactor schedule utilities
