+++ functions/entrar.js
@@ -1,6 +1,7 @@
 import { Redis } from "@upstash/redis";
 import { v4 as uuidv4 } from "uuid";
 import errorHandler from "./utils/errorHandler.js";
+import { withinSchedule } from "./utils/schedule.js";
 
 const LOG_TTL = 60 * 60 * 24 * 30; // 30 days
 
@@ -43,22 +44,6 @@ export async function handler(event) {
         schedule = parsed.schedule || null;
       } catch {}
     }
-
-    const withinSchedule = (sched) => {
-      if (!sched) return true;
-      const tz   = sched.tz || "America/Sao_Paulo";
-      const now  = new Date(new Date().toLocaleString("en-US", { timeZone: tz }));
-      const day  = now.getDay();
-      const days = (sched.days || []).map(Number);
-      if (!days.includes(day)) return false;
-      if (!sched.intervals || sched.intervals.length === 0) return true;
-      const mins = now.getHours() * 60 + now.getMinutes();
-      const toMins = (t) => {
-        const [h, m] = t.split(":").map(Number);
-        return h * 60 + m;
-      };
-      return sched.intervals.some(({ start, end }) => start && end && mins >= toMins(start) && mins < toMins(end));
-    };
 
     // Cria clientId e incrementa contador de tickets
     const clientId     = uuidv4();
