+++ functions/report.js
@@ -1,5 +1,6 @@
 import { Redis } from "@upstash/redis";
 import errorHandler from "./utils/errorHandler.js";
+import { withinSchedule } from "./utils/schedule.js";
 
 export async function handler(event) {
   try {
@@ -93,18 +94,6 @@ export async function handler(event) {
     } else if (monitorRaw) {
       try { schedule = JSON.parse(monitorRaw).schedule || null; } catch {}
     }
-    const withinSchedule = (sched) => {
-      if (!sched) return true;
-      const tz   = sched.tz || "America/Sao_Paulo";
-      const now  = new Date(new Date().toLocaleString("en-US", { timeZone: tz }));
-      const day  = now.getDay();
-      const days = (sched.days || []).map(Number);
-      if (!days.includes(day)) return false;
-      if (!sched.intervals || sched.intervals.length === 0) return true;
-      const mins = now.getHours() * 60 + now.getMinutes();
-      const toMins = (t) => { const [h, m] = t.split(":").map(Number); return h * 60 + m; };
-      return sched.intervals.some(({ start, end }) => start && end && mins >= toMins(start) && mins < toMins(end));
-    };
     if (withinSchedule(schedule)) {
       await redis.srem(prefix + "offHoursSet", ...offHoursSet);
       offHoursSet = [];
