+++ public/client/index.html
@@ -74,6 +74,6 @@
     </div>
   </div>
 
-  <script src="js/client.js"></script>
+  <script type="module" src="js/client.js"></script>
 </body>
 </html>
